move enums to bokeh.core
