+++ bokeh/models/annotations.py
@@ -4,7 +4,7 @@ Bokeh plots
 """
 from __future__ import absolute_import
 
-from ..enums import LegendLocation, SpatialUnits, RenderLevel, Dimension, RenderMode
+from ..core.enums import LegendLocation, SpatialUnits, RenderLevel, Dimension, RenderMode
 from ..property_mixins import LineProps, FillProps, TextProps
 from ..properties import abstract
 from ..properties import (Int, String, Enum, Instance, List, Dict, Tuple,
@@ -46,7 +46,7 @@ class Legend(Annotation):
     location = Either(Enum(LegendLocation), Tuple(Float, Float),
         default="top_right", help="""
     The location where the legend should draw itself. It's either one of
-    ``bokeh.enums.LegendLocation``'s enumerated values, or a ``(x, y)``
+    ``bokeh.core.enums.LegendLocation``'s enumerated values, or a ``(x, y)``
     tuple indicating an absolute location absolute location in screen
     coordinates (pixels from the bottom-left corner).
     """)
