+++ bokeh/models/mappers.py
@@ -6,7 +6,7 @@ from __future__ import absolute_import
 from ..model import Model
 from ..properties import abstract
 from ..properties import Float, Color, Enum, Seq
-from ..enums import Palette
+from ..core.enums import Palette
 from .. import palettes
 
 @abstract
