+++ bokeh/models/renderers.py
@@ -11,7 +11,7 @@ from ..core.validation.errors import BAD_COLUMN_NAME, MISSING_GLYPH, NO_SOURCE_F
 from ..model import Model
 from ..properties import abstract
 from ..properties import String, Enum, Instance, Float, Bool
-from ..enums import RenderLevel
+from ..core.enums import RenderLevel
 
 from .sources import DataSource, RemoteSource
 from .glyphs import Glyph
