+++ bokeh/models/widgets/groups.py
@@ -5,7 +5,7 @@ from __future__ import absolute_import
 
 from ...properties import abstract
 from ...properties import Bool, Int, String, Enum, List
-from ...enums import ButtonType
+from ...core.enums import ButtonType
 from .widget import Widget
 
 @abstract
