+++ bokeh/models/widgets/icons.py
@@ -5,7 +5,7 @@ from __future__ import absolute_import
 
 from ...properties import abstract
 from ...properties import Bool, Float, Enum
-from ...enums import NamedIcon
+from ...core.enums import NamedIcon
 from .widget import Widget
 
 @abstract
